backend/ml-service/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
tensorflow-cpu==2.17.0
numpy==1.26.0
pydantic==2.5.0
python-dotenv
librosa
